i45: Add useMapStorage.js custom hook
