+++ src/library/apis/geolocation.js
@@ -114,16 +114,3 @@ export const stopWatchingLocation = () => {
     store.dispatch({ type: UPDATE_WATCHING, payload: false });
   }
 };
-
-export const onCameraCheck = (userLocation, cameraBounds) => {
-  const isUserLocationWithinCamera = (
-    userLocation[0] < cameraBounds[1][0] ||
-    userLocation[0] > cameraBounds[0][0] ||
-    userLocation[1] > cameraBounds[0][1] ||
-    userLocation[1] < cameraBounds[1][1]);
-
-  if (isUserLocationWithinCamera) {
-    return false;
-  }
-  return true;
-};
